Fix legacy converter for Mul/Add/Sub ops
